requirements.txt
Flask==2.0.3
torch==1.10.2
torchvision==0.11.3
Pillow==8.4.0
numpy==1.21.5
